[apps/reg] Model::numberOfCoefficients return an int, not a double

+++ apps/regression/model/linear_model.h
@@ -15,7 +15,7 @@ public:
   double levelSet(double * modelCoefficients, double y) const override;
   virtual void fit(Store * store, int series, double * modelCoefficients, Poincare::Context * context) override;
   double partialDerivate(double * modelCoefficients, int derivateCoefficientIndex, double x) const override;
-  double numberOfCoefficients() const override { return 2; }
+  int numberOfCoefficients() const override { return 2; }
 };
 
 }
